pkg/proxy/index: move old workspaceindex code here
